+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - new delimiter flag
 - new quote_char flag
 - new header / no-header flags
+- CIDR with host bits set ( = invalid network) translates to IP when asking whois (RIR would return 'invalid search key')
 
 ## 1.1 (2019-05-13)
 - fix: do not reprocess file if moved to another location